@@ -136,7 +136,7 @@ final class xsrlPictureBlockGUI
                 }
                 break;
         }
-        ilUtil::sendFailure($this->plugin->txt('common_access_denied'), true);
+        $this->template->setOnScreenMessage('failure', $this->plugin->txt('common_access_denied'), true);
         $this->controlFlow->redirectByClass(ilRepositoryGUI::class);
 
         return false;
@@ -196,7 +196,7 @@ final class xsrlPictureBlockGUI
                 $anchor .= $block->getSequence();
             }
 
-            ilUtil::sendSuccess($this->plugin->txt('message_changes_save_success'), true);
+            $this->template->setOnScreenMessage('success', $this->plugin->txt('message_changes_save_success'), true);
             $this->controlFlow->redirectByClass(xsrlContentGUI::class, CommonControllerAction::CMD_INDEX, $anchor);
         } catch (ValidationException $ex) {
             $form->setValuesByPost();
@@ -206,7 +206,7 @@ final class xsrlPictureBlockGUI
             $this->template->setContent($form->getHTML());
         } catch (FileUploadException $ex) {
             $form->setValuesByPost();
-            ilUtil::sendFailure($this->plugin->txt('picture_block_upload_error'));
+            $this->template->setOnScreenMessage('failure', $this->plugin->txt('picture_block_upload_error'), true);
             $this->template->setContent($form->getHTML());
         }
     }
@@ -251,7 +251,7 @@ final class xsrlPictureBlockGUI
             $this->pictureBlockService->store($block);
 
             $anchor = xsrlContentGUI::ANCHOR_TEMPLATE . $block->getSequence();
-            ilUtil::sendSuccess($this->plugin->txt('message_changes_save_success'), true);
+            $this->template->setOnScreenMessage('success', $this->plugin->txt('message_changes_save_success'), true);
             $this->controlFlow->redirectByClass(xsrlContentGUI::class, CommonControllerAction::CMD_INDEX, $anchor);
         } catch (ValidationException $ex) {
             $form->setValuesByPost();
@@ -261,7 +261,7 @@ final class xsrlPictureBlockGUI
             $this->template->setContent($form->getHTML());
         } catch (FileUploadException $ex) {
             $form->setValuesByPost();
-            ilUtil::sendFailure($this->plugin->txt('picture_block_upload_error'));
+            $this->template->setOnScreenMessage('failure', $this->plugin->txt('picture_block_upload_error'), true);
             $this->template->setContent($form->getHTML());
         }
     }
@@ -272,7 +272,7 @@ final class xsrlPictureBlockGUI
         $this->redirectInvalidRequests($blockId);
         $this->pictureBlockService->delete($blockId);
         $this->regenerateSequence();
-        ilUtil::sendSuccess($this->plugin->txt('message_delete_success'), true);
+        $this->template->setOnScreenMessage('success', $this->plugin->txt('message_delete_success'), true);
         $this->controlFlow->redirectByClass(xsrlContentGUI::class, CommonControllerAction::CMD_INDEX);
     }
 
