@@ -130,7 +130,7 @@ final class xsrlRichTextBlockGUI
                 }
                 break;
         }
-        ilUtil::sendFailure($this->plugin->txt('common_access_denied'), true);
+        $this->template->setOnScreenMessage('failure', $this->plugin->txt('common_access_denied'), true);
         $this->controlFlow->redirectByClass(ilRepositoryGUI::class);
 
         return false;
@@ -189,7 +189,7 @@ final class xsrlRichTextBlockGUI
                 $anchor .= $block->getSequence();
             }
 
-            ilUtil::sendSuccess($this->plugin->txt('message_changes_save_success'), true);
+            $this->template->setOnScreenMessage('success', $this->plugin->txt('message_changes_save_success'), true);
             $this->controlFlow->redirectByClass(xsrlContentGUI::class, CommonControllerAction::CMD_INDEX, $anchor);
         } catch (ValidationException $ex) {
             $form->setValuesByPost();
@@ -224,7 +224,7 @@ final class xsrlRichTextBlockGUI
             $this->richTextBlockService->store($block);
 
             $anchor = xsrlContentGUI::ANCHOR_TEMPLATE . $block->getSequence();
-            ilUtil::sendSuccess($this->plugin->txt('message_changes_save_success'), true);
+            $this->template->setOnScreenMessage('success', $this->plugin->txt('message_changes_save_success'), true);
             $this->controlFlow->redirectByClass(xsrlContentGUI::class, CommonControllerAction::CMD_INDEX, $anchor);
         } catch (ValidationException $ex) {
             $form->setValuesByPost();
@@ -240,7 +240,7 @@ final class xsrlRichTextBlockGUI
         $this->redirectInvalidRequests($blockId);
         $this->richTextBlockService->delete($blockId);
         $this->regenerateSequence();
-        ilUtil::sendSuccess($this->plugin->txt('message_delete_success'), true);
+        $this->template->setOnScreenMessage('success', $this->plugin->txt('message_delete_success'), true);
         $this->controlFlow->redirectByClass(xsrlContentGUI::class, CommonControllerAction::CMD_INDEX);
     }
 
