@@ -139,7 +139,7 @@ final class xsrlVideoBlockGUI
                 }
                 break;
         }
-        ilUtil::sendFailure($this->plugin->txt('common_access_denied'), true);
+        $this->template->setOnScreenMessage('failure', $this->plugin->txt('common_access_denied'), true);
         $this->controlFlow->redirectByClass(ilRepositoryGUI::class);
 
         return false;
@@ -202,7 +202,7 @@ final class xsrlVideoBlockGUI
                 $anchor .= $block->getSequence();
             }
 
-            ilUtil::sendSuccess($this->plugin->txt('message_changes_save_success'), true);
+            $this->template->setOnScreenMessage('success', $this->plugin->txt('message_changes_save_success'), true);
             $this->controlFlow->redirectByClass(xsrlContentGUI::class, CommonControllerAction::CMD_INDEX, $anchor);
         } catch (ValidationException $ex) {
             $form->setValuesByPost();
@@ -212,7 +212,7 @@ final class xsrlVideoBlockGUI
             $this->template->setContent($form->getHTML());
         } catch (FileUploadException $ex) {
             $form->setValuesByPost();
-            ilUtil::sendFailure($this->plugin->txt('video_block_upload_error'));
+            $this->template->setOnScreenMessage('failure', $this->plugin->txt('video_block_upload_error'), true);
             $this->template->setContent($form->getHTML());
         }
     }
@@ -263,7 +263,7 @@ final class xsrlVideoBlockGUI
             $this->videoBlockService->store($block);
 
             $anchor = xsrlContentGUI::ANCHOR_TEMPLATE . $block->getSequence();
-            ilUtil::sendSuccess($this->plugin->txt('message_changes_save_success'), true);
+            $this->template->setOnScreenMessage('success', $this->plugin->txt('message_changes_save_success'), true);
             $this->controlFlow->redirectByClass(xsrlContentGUI::class, CommonControllerAction::CMD_INDEX, $anchor);
         } catch (ValidationException $ex) {
             $form->setValuesByPost();
@@ -273,7 +273,7 @@ final class xsrlVideoBlockGUI
             $this->template->setContent($form->getHTML());
         } catch (FileUploadException $ex) {
             $form->setValuesByPost();
-            ilUtil::sendFailure($this->plugin->txt('video_block_upload_error'));
+            $this->template->setOnScreenMessage('failure', $this->plugin->txt('video_block_upload_error'), true);
             $this->template->setContent($form->getHTML());
         }
     }
@@ -285,7 +285,7 @@ final class xsrlVideoBlockGUI
         $this->redirectInvalidRequests($blockId);
         $this->videoBlockService->delete($blockId);
         $this->regenerateSequence();
-        ilUtil::sendSuccess($this->plugin->txt('message_delete_success'), true);
+        $this->template->setOnScreenMessage('success', $this->plugin->txt('message_delete_success'), true);
         $this->controlFlow->redirectByClass(xsrlContentGUI::class, CommonControllerAction::CMD_INDEX);
     }
 
