@@ -1,7 +1,7 @@
 <?php
 $id = "xsrl";
 $version = "4.0.2";
-$ilias_min_version = "6.0";
-$ilias_max_version = "7.99";
+$ilias_min_version = "8.0";
+$ilias_max_version = "8.99";
 $responsible = "fluxlabs AG";
 $responsible_mail = "support@fluxlabs.ch";
